+++ src/components/text-editor/__internal__/__plugins__/index.ts
@@ -2,3 +2,4 @@ export { default as AutoLinkerPlugin } from "./AutoLinker";
 export { default as LinkMonitorPlugin } from "./LinkMonitor";
 export { default as StyledSpanEnterPlugin } from "./StyledSpanEnter";
 export { default as useCursorAtEnd } from "./useCursorAtEnd";
+export { default as FormSubmissionPlugin } from "./FormSubmission";